+++ src/main/java/org/webbitserver/stub/StubConnection.java
@@ -52,7 +52,14 @@ public class StubConnection extends StubDataHolder implements EventSourceConnect
 
     @Override
     public StubConnection send(byte[] message) {
-        sentBinaryMessages.add(message);
+        return send(message, 0, message.length);
+    }
+
+    @Override
+    public StubConnection send(byte[] message, int offset, int length) {
+        byte[] subMessage = new byte[length];
+        System.arraycopy(message, offset, subMessage, 0, length);
+        sentBinaryMessages.add(subMessage);
         return this;
     }